requirements.txt
numpy==1.24.3
Pillow==9.5.0
pandas==1.5.3
matplotlib==3.7.1
scipy